@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@material-ui/core": "^1.2.0",
     "autoprefixer": "7.1.6",
     "babel-core": "6.26.0",
     "babel-eslint": "7.2.3",
@@ -12,7 +13,7 @@
     "babel-runtime": "6.26.0",
     "case-sensitive-paths-webpack-plugin": "2.1.1",
     "chalk": "1.1.3",
-    "classnames": "^2.2.5",
+    "classnames": "^2.2.6",
     "css-loader": "0.28.7",
     "date-fns": "2.0.0-alpha.7",
     "dotenv": "4.0.0",
@@ -26,30 +27,29 @@
     "history": "^4.7.2",
     "html-webpack-plugin": "2.29.0",
     "jest": "20.0.4",
-    "jss": "^9.8.1",
-    "jss-preset-default": "^4.3.0",
+    "jss": "^9.8.2",
+    "jss-preset-default": "^4.5.0",
     "jss-rtl": "^0.2.3",
-    "luxon": "^1.2.0",
-    "material-ui": "^1.0.0-beta.43",
-    "material-ui-pickers-jalali-utils": "^0.4.0",
-    "moment": "^2.22.1",
-    "moment-jalaali": "^0.7.2",
+    "luxon": "^1.2.1",
+    "material-ui-pickers-jalali-utils": "^0.4.2",
+    "moment": "^2.22.2",
+    "moment-jalaali": "^0.7.4",
     "object-assign": "4.1.1",
-    "patreon": "^0.4.0",
+    "patreon": "^0.4.1",
     "postcss-flexbugs-fixes": "3.2.0",
     "postcss-loader": "2.0.8",
     "prismjs": "^1.14.0",
     "promise": "8.0.1",
     "prop-types": "^15.6.1",
     "raf": "3.4.0",
-    "react": "^16.3.2",
+    "react": "^16.4.0",
     "react-dev-utils": "^5.0.1",
     "react-docgen": "^2.20.1",
-    "react-dom": "^16.3.2",
-    "react-jss": "^8.4.0",
+    "react-dom": "^16.4.0",
+    "react-jss": "^8.5.0",
     "react-markdown": "^2.5.1",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
     "style-loader": "0.19.0",
     "sw-precache-webpack-plugin": "0.11.4",
     "url-loader": "0.6.2",
@@ -59,13 +59,13 @@
     "whatwg-fetch": "2.0.3"
   },
   "devDependencies": {
-    "cross-env": "^5.1.4",
+    "cross-env": "^5.1.6",
     "eslint": "^4.19.1",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-import-resolver-webpack": "^0.8.4",
-    "eslint-plugin-import": "^2.11.0",
+    "eslint-plugin-import": "^2.12.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.7.0",
+    "eslint-plugin-react": "^7.9.1",
     "raw-loader": "^0.5.1"
   },
   "scripts": {
