@@ -1,8 +1,8 @@
 import React, { PureComponent } from 'react';
 import { DatePicker } from 'material-ui-pickers';
-import { MuiThemeProvider, createMuiTheme } from 'material-ui';
+import { MuiThemeProvider, createMuiTheme } from '@material-ui/core';
 
-import lightBlue from 'material-ui/colors/lightBlue';
+import lightBlue from '@material-ui/core/colors/lightBlue';
 
 const materialTheme = createMuiTheme({
   overrides: {
