@@ -1,6 +1,6 @@
 import React from 'react';
 import SourcablePanel from '_shared/SourcablePanel';
-import { Typography } from 'material-ui';
+import { Typography } from '@material-ui/core';
 
 const ControllingProgrammatically = () => (
   <div>
