@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { ListItem, withStyles } from 'material-ui';
-import Collapse from 'material-ui/transitions/Collapse';
-import Button from 'material-ui/Button';
+import { ListItem, withStyles, Collapse } from '@material-ui/core';
+
+import Button from '@material-ui/core/Button';
 import { NavLink } from 'react-router-dom';
 
 const styles = theme => ({
