@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { List } from 'material-ui';
+import { List } from '@material-ui/core';
 import { withRouter } from 'react-router-dom';
 import NavItem from './NavItem';
 
