@@ -27,6 +27,7 @@ export class Calendar extends Component {
     theme: PropTypes.object.isRequired,
     shouldDisableDate: PropTypes.func,
     utils: PropTypes.object.isRequired,
+    allowKeyboardControl: PropTypes.bool,
   };
 
   static defaultProps = {
@@ -37,8 +38,20 @@ export class Calendar extends Component {
     leftArrowIcon: undefined,
     rightArrowIcon: undefined,
     renderDay: undefined,
+    allowKeyboardControl: false,
     shouldDisableDate: () => false,
   };
+
+  static getDerivedStateFromProps(nextProps, state) {
+    if (!nextProps.utils.isEqual(nextProps.date, state.lastDate)) {
+      return {
+        lastDate: nextProps.date,
+        currentMonth: nextProps.utils.getStartOfMonth(nextProps.date),
+      };
+    }
+
+    return null;
+  }
 
   state = {
     currentMonth: this.props.utils.getStartOfMonth(this.props.date),
@@ -60,12 +73,6 @@ export class Calendar extends Component {
         shouldDisableDate: this.shouldDisableDate,
       }));
     }
-  }
-
-  componentWillReceiveProps(nextProps) {
-    this.setState({
-      currentMonth: this.props.utils.getStartOfMonth(nextProps.date),
-    });
   }
 
   onDateSelect = (day) => {
@@ -198,12 +205,7 @@ export class Calendar extends Component {
       );
 
       if (renderDay) {
-        dayComponent = renderDay(
-          day,
-          selectedDate,
-          dayInCurrentMonth,
-          dayComponent,
-        );
+        dayComponent = renderDay(day, selectedDate, dayInCurrentMonth, dayComponent);
       }
 
       return (
@@ -222,11 +224,14 @@ export class Calendar extends Component {
 
   render() {
     const { currentMonth } = this.state;
-    const { classes, utils } = this.props;
+    const { classes, utils, allowKeyboardControl } = this.props;
 
     return (
       <Fragment>
-        <EventListener target="window" onKeyDown={this.handleKeyDown} />
+        {
+          allowKeyboardControl &&
+            <EventListener target="window" onKeyDown={this.handleKeyDown} />
+        }
 
         <CalendarHeader
           currentMonth={currentMonth}
@@ -238,7 +243,12 @@ export class Calendar extends Component {
           utils={utils}
         />
 
-        <div className={classes.calendar}>{this.renderWeeks()}</div>
+        <div
+          autoFocus /* eslint-disable-line */ // Autofocus required for getting work keyboard navigation feature
+          className={classes.calendar}
+        >
+          {this.renderWeeks()}
+        </div>
       </Fragment>
     );
   }
