@@ -4,22 +4,20 @@ import { Utils } from '../typings/utils';
 import { RenderDay } from './Calendar';
 import { ModalWrapperProps } from '../wrappers/ModalWrapper';
 import { MaterialUiPickersDate } from '../typings/date'
-import { PickerBaseProps } from '../_shared/PickerBase'
+import { BasePickerProps } from '../_shared/BasePicker'
 import { Omit } from '@material-ui/core'
 
-export interface DatePickerWrapperProps extends PickerBaseProps,
-  Omit<ModalWrapperProps, 'onChange'> {
-    minDate?: DateType;
-    maxDate?: DateType;
-    disablePast?: boolean;
-    disableFuture?: boolean;
-    animateYearScrolling?: boolean;
-    openToYearSelection?: boolean;
-    leftArrowIcon?: ReactNode;
-    rightArrowIcon?: ReactNode;
-    renderDay?: RenderDay;
-    utils?: Utils<MaterialUiPickersDate>;
-    shouldDisableDate?: (day: MaterialUiPickersDate) => boolean;
+export interface DatePickerWrapperProps extends BasePickerProps, Omit<ModalWrapperProps, 'onChange' | 'value'> {
+  minDate?: DateType;
+  maxDate?: DateType;
+  disablePast?: boolean;
+  disableFuture?: boolean;
+  animateYearScrolling?: boolean;
+  openToYearSelection?: boolean;
+  leftArrowIcon?: ReactNode;
+  rightArrowIcon?: ReactNode;
+  renderDay?: RenderDay;
+  shouldDisableDate?: (day: MaterialUiPickersDate) => boolean;
 }
 
 declare const DatePickerWrapper: ComponentClass<DatePickerWrapperProps>;
