@@ -4,123 +4,137 @@ import PropTypes from 'prop-types';
 import ModalWrapper from '../wrappers/ModalWrapper';
 import DatePicker from './DatePicker';
 import DomainPropTypes from '../constants/prop-types';
-import PickerBase from '../_shared/PickerBase';
+import BasePicker from '../_shared/BasePicker';
 import withUtils from '../_shared/WithUtils';
 
-export class DatePickerWrapper extends PickerBase {
-  static propTypes = {
-    utils: PropTypes.object.isRequired,
-    /** Datepicker value */
-    value: DomainPropTypes.date,
-    /** Min selectable date */
-    minDate: DomainPropTypes.date,
-    /** Max selectable date */
-    maxDate: DomainPropTypes.date,
-    /** Date format string for input */
-    format: PropTypes.string,
-    /** Callback firing when date accepted */
-    onChange: PropTypes.func.isRequired,
-    /** Auto accept date on selection */
-    autoOk: PropTypes.bool,
-    /** Disable past dates */
-    disablePast: PropTypes.bool,
-    /** Disable future dates */
-    disableFuture: PropTypes.bool,
-    /** To animate scrolling to current year (with scrollIntoView) */
-    animateYearScrolling: PropTypes.bool,
-    /** Open datepicker from year selection */
-    openToYearSelection: PropTypes.bool,
-    /** Allow to specify dynamic label for text field labelFunc(date, invalidLabel) */
-    labelFunc: PropTypes.func,
-    /** Left arrow icon */
-    leftArrowIcon: PropTypes.node,
-    /** Right arrow icon */
-    rightArrowIcon: PropTypes.node,
-    /** Custom renderer for day renderDay(date, selectedDate, dayInCurrentMonth) */
-    renderDay: PropTypes.func,
-    /** Disable specific date */
-    shouldDisableDate: PropTypes.func,
-  }
+export const DatePickerWrapper = (props) => {
+  const {
+    allowKeyboardControl,
+    animateYearScrolling,
+    autoOk,
+    disableFuture,
+    disablePast,
+    format,
+    forwardedRef,
+    labelFunc,
+    leftArrowIcon,
+    maxDate,
+    minDate,
+    onChange,
+    openToYearSelection,
+    renderDay,
+    rightArrowIcon,
+    shouldDisableDate,
+    utils,
+    value,
+    ...other
+  } = props;
 
-  static defaultProps = {
-    value: new Date(),
-    format: 'MMMM Do',
-    autoOk: false,
-    minDate: '1900-01-01',
-    maxDate: '2100-01-01',
-    disableFuture: false,
-    disablePast: false,
-    animateYearScrolling: false,
-    openToYearSelection: false,
-    leftArrowIcon: 'keyboard_arrow_left',
-    rightArrowIcon: 'keyboard_arrow_right',
-    renderDay: undefined,
-    labelFunc: undefined,
-    shouldDisableDate: undefined,
-  }
+  return (
+    <BasePicker {...props}>
+      {
+        ({
+          date,
+          handleAccept,
+          handleChange,
+          handleClear,
+          handleDismiss,
+          handleSetTodayDate,
+          handleTextFieldChange,
+        }) => (
+          <ModalWrapper
+            disableFuture={disableFuture}
+            disablePast={disablePast}
+            format={format}
+            labelFunc={labelFunc}
+            maxDate={maxDate}
+            minDate={minDate}
+            onAccept={handleAccept}
+            onChange={handleTextFieldChange}
+            onClear={handleClear}
+            onDismiss={handleDismiss}
+            onSetToday={handleSetTodayDate}
+            {...other}
+          >
+            <DatePicker
+              allowKeyboardControl={allowKeyboardControl}
+              animateYearScrolling={animateYearScrolling}
+              date={date}
+              disableFuture={disableFuture}
+              disablePast={disablePast}
+              leftArrowIcon={leftArrowIcon}
+              maxDate={maxDate}
+              minDate={minDate}
+              onChange={handleChange}
+              openToYearSelection={openToYearSelection}
+              renderDay={renderDay}
+              rightArrowIcon={rightArrowIcon}
+              shouldDisableDate={shouldDisableDate}
+              utils={utils}
+            />
+          </ModalWrapper>
+        )
+      }
+    </BasePicker>
+  );
+};
 
-  render() {
-    const { date } = this.state;
-    const {
-      value,
-      format,
-      autoOk,
-      onChange,
-      animateYearScrolling,
-      openToYearSelection,
-      leftArrowIcon,
-      rightArrowIcon,
-      renderDay,
-      labelFunc,
-      utils,
-      shouldDisableDate,
-      minDateMessage,
-      maxDateMessage,
-      minDate,
-      maxDate,
-      disablePast,
-      disableFuture,
-      ...other
-    } = this.props;
+DatePickerWrapper.propTypes = {
+  utils: PropTypes.object.isRequired,
+  /** Datepicker value */
+  value: DomainPropTypes.date,
+  /** Min selectable date */
+  minDate: DomainPropTypes.date,
+  /** Max selectable date */
+  maxDate: DomainPropTypes.date,
+  /** Date format string for input */
+  format: PropTypes.string,
+  /** Callback firing when date accepted */
+  onChange: PropTypes.func.isRequired,
+  /** Auto accept date on selection */
+  autoOk: PropTypes.bool,
+  /** Disable past dates */
+  disablePast: PropTypes.bool,
+  /** Disable future dates */
+  disableFuture: PropTypes.bool,
+  /** To animate scrolling to current year (with scrollIntoView) */
+  animateYearScrolling: PropTypes.bool,
+  /** Open datepicker from year selection */
+  openToYearSelection: PropTypes.bool,
+  /** Allow to specify dynamic label for text field labelFunc(date, invalidLabel) */
+  labelFunc: PropTypes.func,
+  /** Left arrow icon */
+  leftArrowIcon: PropTypes.node,
+  /** Right arrow icon */
+  rightArrowIcon: PropTypes.node,
+  /** Custom renderer for day renderDay(date, selectedDate, dayInCurrentMonth) */
+  renderDay: PropTypes.func,
+  /** Disable specific date */
+  shouldDisableDate: PropTypes.func,
+  /** Enables keyboard listener for moving between days in calendar */
+  allowKeyboardControl: PropTypes.bool,
+  forwardedRef: PropTypes.func,
+};
 
-    return (
-      <ModalWrapper
-        ref={this.getRef}
-        value={value}
-        format={format}
-        onClear={this.handleClear}
-        onAccept={this.handleAccept}
-        onChange={this.handleTextFieldChange}
-        onDismiss={this.handleDismiss}
-        onSetToday={this.handleSetTodayDate}
-        labelFunc={labelFunc}
-        minDate={minDate}
-        maxDate={maxDate}
-        disablePast={disablePast}
-        disableFuture={disableFuture}
-        minDateMessage={minDateMessage}
-        maxDateMessage={maxDateMessage}
-        {...other}
-      >
-        <DatePicker
-          date={date}
-          onChange={this.handleChange}
-          animateYearScrolling={animateYearScrolling}
-          openToYearSelection={openToYearSelection}
-          leftArrowIcon={leftArrowIcon}
-          rightArrowIcon={rightArrowIcon}
-          renderDay={renderDay}
-          utils={utils}
-          minDate={minDate}
-          maxDate={maxDate}
-          disablePast={disablePast}
-          disableFuture={disableFuture}
-          shouldDisableDate={shouldDisableDate}
-        />
-      </ModalWrapper>
-    );
-  }
-}
+DatePickerWrapper.defaultProps = {
+  value: new Date(),
+  format: 'MMMM Do',
+  autoOk: false,
+  minDate: '1900-01-01',
+  maxDate: '2100-01-01',
+  disableFuture: false,
+  disablePast: false,
+  animateYearScrolling: false,
+  openToYearSelection: false,
+  allowKeyboardControl: true,
+  leftArrowIcon: 'keyboard_arrow_left',
+  rightArrowIcon: 'keyboard_arrow_right',
+  renderDay: undefined,
+  labelFunc: undefined,
+  shouldDisableDate: undefined,
+  forwardedRef: undefined,
+};
 
-export default withUtils()(DatePickerWrapper);
+const WithUtils = withUtils()(DatePickerWrapper);
+export default React.forwardRef((props, ref) => <WithUtils {...props} forwardedRef={ref} />);
 
