@@ -17,41 +17,43 @@ import withUtils from '../_shared/WithUtils';
 
 export class DateTimePicker extends Component {
   static propTypes = {
-    date: PropTypes.object.isRequired,
-    onChange: PropTypes.func.isRequired,
-    autoSubmit: PropTypes.bool,
-    openTo: PropTypes.oneOf(Object.keys(viewType).map(key => viewType[key])).isRequired,
-    disablePast: PropTypes.bool,
-    disableFuture: PropTypes.bool,
-    minDate: DomainPropTypes.date.isRequired,
-    maxDate: DomainPropTypes.date.isRequired,
-    showTabs: PropTypes.bool,
-    leftArrowIcon: PropTypes.node,
-    rightArrowIcon: PropTypes.node,
-    dateRangeIcon: PropTypes.node,
-    timeIcon: PropTypes.node,
-    renderDay: PropTypes.func,
-    utils: PropTypes.object.isRequired,
+    allowKeyboardControl: PropTypes.bool,
     ampm: PropTypes.bool,
-    shouldDisableDate: PropTypes.func,
     animateYearScrolling: PropTypes.bool,
-    fadeTimeout: PropTypes.number.isRequired,
+    autoSubmit: PropTypes.bool,
     classes: PropTypes.object.isRequired,
+    date: PropTypes.object.isRequired,
+    dateRangeIcon: PropTypes.node,
+    disableFuture: PropTypes.bool,
+    disablePast: PropTypes.bool,
+    fadeTimeout: PropTypes.number.isRequired,
+    leftArrowIcon: PropTypes.node,
+    maxDate: DomainPropTypes.date.isRequired,
+    minDate: DomainPropTypes.date.isRequired,
+    onChange: PropTypes.func.isRequired,
+    openTo: PropTypes.oneOf(Object.keys(viewType).map(key => viewType[key])).isRequired,
+    renderDay: PropTypes.func,
+    rightArrowIcon: PropTypes.node,
+    shouldDisableDate: PropTypes.func,
+    showTabs: PropTypes.bool,
+    timeIcon: PropTypes.node,
+    utils: PropTypes.object.isRequired,
   }
 
   static defaultProps = {
-    autoSubmit: true,
-    disablePast: false,
-    disableFuture: false,
-    showTabs: true,
-    leftArrowIcon: undefined,
-    rightArrowIcon: undefined,
-    dateRangeIcon: undefined,
-    timeIcon: undefined,
-    renderDay: undefined,
+    allowKeyboardControl: false,
     ampm: true,
-    shouldDisableDate: undefined,
     animateYearScrolling: false,
+    autoSubmit: true,
+    dateRangeIcon: undefined,
+    disableFuture: false,
+    disablePast: false,
+    leftArrowIcon: undefined,
+    renderDay: undefined,
+    rightArrowIcon: undefined,
+    shouldDisableDate: undefined,
+    showTabs: true,
+    timeIcon: undefined,
   }
 
   state = {
@@ -120,6 +122,7 @@ export class DateTimePicker extends Component {
       animateYearScrolling,
       fadeTimeout,
       classes,
+      allowKeyboardControl,
     } = this.props;
 
     return (
@@ -160,6 +163,7 @@ export class DateTimePicker extends Component {
 
           <View view={viewType.DATE} selected={openView}>
             <Calendar
+              allowKeyboardControl={allowKeyboardControl}
               date={date}
               minDate={minDate}
               maxDate={maxDate}
