@@ -1,13 +1,12 @@
 import { ComponentClass } from 'react';
 import { Utils } from '../typings/utils';
 import { ModalWrapperProps } from '../wrappers/ModalWrapper';
-import { PickerBaseProps } from '../_shared/PickerBase'
+import { BasePickerProps } from '../_shared/BasePicker'
 import { Omit } from '@material-ui/core'
 import { MaterialUiPickersDate } from '../typings/date';
 
-export interface TimePickerWrapperProps extends PickerBaseProps,
-  Omit<ModalWrapperProps, 'onChange'> {
-    utils?: Utils<MaterialUiPickersDate>;
+export interface TimePickerWrapperProps extends BasePickerProps, Omit<ModalWrapperProps, 'onChange' | 'value'> {
+
 }
 
 declare const TimePickerWrapper: ComponentClass<TimePickerWrapperProps>;
