@@ -121,11 +121,11 @@ ModalDialog.propTypes = {
   onClear: PropTypes.func.isRequired,
   classes: PropTypes.object.isRequired,
   dialogContentClassName: PropTypes.string,
-  okLabel: PropTypes.string.isRequired,
-  cancelLabel: PropTypes.string.isRequired,
-  clearLabel: PropTypes.string.isRequired,
+  okLabel: PropTypes.node.isRequired,
+  cancelLabel: PropTypes.node.isRequired,
+  clearLabel: PropTypes.node.isRequired,
   clearable: PropTypes.bool.isRequired,
-  todayLabel: PropTypes.string.isRequired,
+  todayLabel: PropTypes.node.isRequired,
   showTodayButton: PropTypes.bool.isRequired,
   onSetToday: PropTypes.func.isRequired,
 };
